studio or changes merge

+++ web_module/trunk/src/main/com/topcoder/web/common/dao/hibernate/ProjectDAOHibernate.java
@@ -1,11 +1,10 @@
 package com.topcoder.web.common.dao.hibernate;
 
-import java.util.List;
-
-import org.hibernate.Session;
-
 import com.topcoder.web.common.dao.ProjectDAO;
 import com.topcoder.web.common.model.comp.Project;
+import org.hibernate.Session;
+
+import java.util.List;
 
 /**
  * @author cucu
@@ -20,7 +19,7 @@ public class ProjectDAOHibernate extends Base implements ProjectDAO {
         super(session);
     }
 
-    public Project find(Long id) {        
+    public Project find(Integer id) {
         return (Project) find(Project.class, id);
     }
 
@@ -28,8 +27,8 @@ public class ProjectDAOHibernate extends Base implements ProjectDAO {
     public List<Project> find(Integer statusId, Integer categoryId) {
         return session.createQuery(
                 "FROM com.topcoder.web.common.model.comp.Project p  " +
-                "WHERE p.statusId = :statusId " +
-                "AND p.categoryId = :categoryId")
+                        "WHERE p.statusId = :statusId " +
+                        "AND p.categoryId = :categoryId")
                 .setInteger("statusId", statusId)
                 .setInteger("categoryId", categoryId)
                 .list();
